refact(examples + Thread) some renaming + no bundle builded
