+++ examples/Window.html
@@ -4,7 +4,7 @@
     <title>Window example</title>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./assets/css/default.css" />
+    <link rel="stylesheet" href="./assets/css/examples.css" />
   </head>
   <body>
     <script src="../packages/browser/dist/RUN_MODE/bundle.js"></script>
