+++ packages/node/src/Game/Game.js
@@ -2,5 +2,4 @@ module.exports = {
   Thread: require('./Thread'),
   ScriptTemplate: require('./ScriptTemplate/ScriptTemplate'),
   SocketService: require('./SocketService'),
-  ThreadProcessRoutine: require('./ThreadProcessRoutine'),
 };
