+++ packages/node/src/Game/Thread.js
@@ -1,12 +1,43 @@
 const workerThreads = require('worker_threads');
-const { Data, Constant } = require('@ud-viz/shared');
+const {
+  Data,
+  Constant,
+  Game,
+  ProcessInterval,
+  Command,
+} = require('@ud-viz/shared');
 const SocketWrapper = require('./SocketWrapper');
 const THREE = require('three');
 
 /**
+ * Different Event between Parent and Child
+ */
+const EVENT = {
+  // parent => child
+  INIT: 'init',
+  COMMANDS: 'commands',
+  ADD_OBJECT3D: 'add_object3D',
+  REMOVE_OBJECT3D: 'remove_object3D',
+  ON_NEW_SOCKET_WRAPPER: 'on_new_socket_wrapper',
+  ON_SOCKET_WRAPPER_REMOVE: 'on_socket_wrapper_remove',
+  // child => parent
+  CURRENT_STATE: 'current_state',
+  APPLY_RESOLVE: 'apply_resolve',
+};
+
+/**
+ * Event message structure
+ */
+const KEY = {
+  DATA: 0,
+  TYPE: 1,
+  APPLY_UUID: 2, // mean parent is waiting an answer from child to resolve apply promise
+};
+
+/**
  * @classdesc - {@link workerThreads} wrapper, different event can be send/receive by the thread
  */
-const Thread = class {
+const Parent = class {
   /**
    * Manage communication between socket wrapper and worker thread
    *
@@ -36,15 +67,13 @@ const Thread = class {
     // listen
     this.worker.on('message', (int32ArrayMessage) => {
       const objectMessage = Data.int32ArrayToObject(int32ArrayMessage);
-      if (this.callbacks[objectMessage[Thread.KEY.TYPE]]) {
-        this.callbacks[objectMessage[Thread.KEY.TYPE]](
-          objectMessage[Thread.KEY.DATA]
-        );
+      if (this.callbacks[objectMessage[KEY.TYPE]]) {
+        this.callbacks[objectMessage[KEY.TYPE]](objectMessage[KEY.DATA]);
       }
     });
 
     // wait resolve signal from child thread
-    this.on(Thread.EVENT.APPLY_RESOLVE, (applyUUID) => {
+    this.on(EVENT.APPLY_RESOLVE, (applyUUID) => {
       const resolve = this._resolveApply[applyUUID];
       if (!resolve) throw new Error('no resolve for ', applyUUID);
       resolve();
@@ -59,7 +88,7 @@ const Thread = class {
    */
   addSocketWrapper(socketWrapper) {
     this.socketWrappers.push(socketWrapper);
-    this.post(Thread.EVENT.ON_NEW_SOCKET_WRAPPER, socketWrapper.socket.id);
+    this.post(EVENT.ON_NEW_SOCKET_WRAPPER, socketWrapper.socket.id);
 
     // reset last state of socket wrapper
     socketWrapper.lastStateSend = null;
@@ -71,7 +100,7 @@ const Thread = class {
     socketWrapper.socket.on(
       Constant.WEBSOCKET.MSG_TYPE.COMMANDS,
       (commands) => {
-        this.post(Thread.EVENT.COMMANDS, commands);
+        this.post(EVENT.COMMANDS, commands);
       }
     );
   }
@@ -84,7 +113,7 @@ const Thread = class {
   removeSocketWrapper(socketWrapper) {
     const index = this.socketWrappers.indexOf(socketWrapper);
     if (index >= 0) this.socketWrappers.splice(index, 1);
-    this.post(Thread.EVENT.ON_SOCKET_WRAPPER_REMOVE, socketWrapper.socket.id);
+    this.post(EVENT.ON_SOCKET_WRAPPER_REMOVE, socketWrapper.socket.id);
     socketWrapper.socket.removeAllListeners(
       Constant.WEBSOCKET.MSG_TYPE.COMMANDS
     );
@@ -100,8 +129,8 @@ const Thread = class {
    */
   post(msgType, data) {
     const object = {};
-    object[Thread.KEY.TYPE] = msgType;
-    object[Thread.KEY.DATA] = data;
+    object[KEY.TYPE] = msgType;
+    object[KEY.DATA] = data;
     this.worker.postMessage(Data.objectToInt32Array(object));
   }
 
@@ -126,10 +155,10 @@ const Thread = class {
   apply(msgType, data) {
     return new Promise((resolve) => {
       const object = {};
-      object[Thread.KEY.TYPE] = msgType;
-      object[Thread.KEY.DATA] = data;
+      object[KEY.TYPE] = msgType;
+      object[KEY.DATA] = data;
       const applyUUID = THREE.MathUtils.generateUUID();
-      object[Thread.KEY.APPLY_UUID] = applyUUID;
+      object[KEY.APPLY_UUID] = applyUUID;
       this.worker.postMessage(Data.objectToInt32Array(object));
 
       this._resolveApply[applyUUID] = resolve;
@@ -137,29 +166,178 @@ const Thread = class {
   }
 };
 
-/**
- * Different Event between Thread and workerThread.Worker
- */
-Thread.EVENT = {
-  // parent => child
-  INIT: 'init',
-  COMMANDS: 'commands',
-  ADD_OBJECT3D: 'add_object3D',
-  REMOVE_OBJECT3D: 'remove_object3D',
-  ON_NEW_SOCKET_WRAPPER: 'on_new_socket_wrapper',
-  ON_SOCKET_WRAPPER_REMOVE: 'on_socket_wrapper_remove',
-  // child => parent
-  CURRENT_STATE: 'current_state',
-  APPLY_RESOLVE: 'apply_resolve',
-};
+/** Code below is used in the thread child */
 
 /**
- * Message structure
+ *
+ * @param {Object<string,Function>} gameScriptClass - class needs by object3D
+ * @returns {Promise} - a promise resolving when game context is initialized and returning thread context {@link Child}
  */
-Thread.KEY = {
-  DATA: 0,
-  TYPE: 1,
-  APPLY_UUID: 2, // mean main thread is waiting an answer from child thread to resolve apply promise
-};
+function childProcess(gameScriptClass = {}) {
+  return new Promise((resolve) => {
+    if (workerThreads.isMainThread) {
+      throw new Error('Its not a worker');
+    }
 
-module.exports = Thread;
+    const parentPort = workerThreads.parentPort;
+
+    const threadChild = new Child(parentPort);
+
+    /** @type {Game.Context} */
+    let gameContext = null;
+    let commands = null;
+
+    parentPort.on('message', (int32ArrayMessage) => {
+      const objectMessage = Data.int32ArrayToObject(int32ArrayMessage);
+      const data = objectMessage[KEY.DATA];
+      const applyUUID = objectMessage[KEY.APPLY_UUID];
+      const promises = [];
+
+      // dispatch for custom event & record promise associated for apply resolve
+      promises.push(threadChild.dispatch(objectMessage[KEY.TYPE], data));
+
+      switch (objectMessage[KEY.TYPE]) {
+        case EVENT.INIT:
+          gameContext = new Game.Context(
+            gameScriptClass,
+            new Game.Object3D(data.gameObject3D)
+          );
+
+          // init is not sync record promise
+          promises.push(
+            gameContext.load().then(() => {
+              const process = new ProcessInterval({ fps: 60 });
+              process.start((dt) => {
+                // simulation
+                gameContext.step(dt);
+                const currentState = gameContext.toState(false); // false because no need to send component already controlled
+                // post state to main thread
+                const message = {};
+                message[KEY.TYPE] = EVENT.CURRENT_STATE;
+                message[KEY.DATA] = currentState.toJSON();
+                parentPort.postMessage(Data.objectToInt32Array(message));
+              });
+
+              console.log(
+                'Child process ',
+                gameContext.object3D.name,
+                ' initialized',
+                gameContext.object3D.uuid
+              );
+
+              threadChild.initGameContext(gameContext);
+
+              resolve(threadChild);
+            })
+          );
+          break;
+        case EVENT.COMMANDS:
+          commands = [];
+          data.forEach(function (c) {
+            commands.push(new Command(c));
+          });
+          gameContext.onCommands(commands);
+          break;
+        case EVENT.ADD_OBJECT3D:
+          // add is not sync record in promises for apply
+          promises.push(
+            gameContext.addObject3D(
+              new Game.Object3D(data.object3D),
+              data.parentUUID
+            )
+          );
+          break;
+        case EVENT.REMOVE_OBJECT3D:
+          gameContext.removeObject3D(data);
+          break;
+        case EVENT.ON_NEW_SOCKET_WRAPPER:
+          gameContext.dispatch(EVENT.ON_NEW_SOCKET_WRAPPER, data);
+          break;
+        case EVENT.ON_SOCKET_WRAPPER_REMOVE:
+          gameContext.dispatch(EVENT.ON_SOCKET_WRAPPER_REMOVE, data);
+          break;
+        default:
+          console.warn(objectMessage[KEY.TYPE], ' not handle natively');
+      }
+
+      if (applyUUID) {
+        Promise.all(promises).then(() => {
+          const applyResolveMessage = {};
+          applyResolveMessage[KEY.TYPE] = EVENT.APPLY_RESOLVE;
+          applyResolveMessage[KEY.DATA] = applyUUID;
+          parentPort.postMessage(Data.objectToInt32Array(applyResolveMessage));
+        });
+      }
+    });
+  });
+}
+
+/**
+ * @class class containing all information to manipulate a worker thread
+ */
+class Child {
+  /**
+   *
+   * @param {workerThreads.MessagePort} parentPort - parent port of this thread
+   */
+  constructor(parentPort) {
+    /** @type {Game.Context} */
+    this.gameContext = null;
+
+    /** @type {workerThreads.MessagePort} */
+    this.parentPort = parentPort;
+
+    /** @type {Object<string,Promise>} */
+    this.promises = {};
+  }
+
+  /**
+   *
+   * @param {Game.Context} value - game context of thread
+   */
+  initGameContext(value) {
+    this.gameContext = value;
+  }
+
+  /**
+   * @callback PromiseListener
+   * @param {*} params
+   * @returns {Promise}
+   */
+
+  /**
+   * Add a listener to parent thread message
+   *
+   * @param {string} eventID - event id
+   * @param {PromiseListener|undefined} promise - callback return a promise or nothing
+   */
+  on(eventID, promise) {
+    for (const event in Parent.EVENT) {
+      if (EVENT[event] === eventID) {
+        throw new Error('native Thread event');
+      }
+    }
+
+    this.promises[eventID] = promise;
+  }
+
+  /**
+   *
+   * @param {string} eventID - event to notify listener of
+   * @param {object} data - serializable data
+   * @returns {Promise} - promise resolving when event has finished
+   */
+  dispatch(eventID, data) {
+    if (this.promises[eventID]) {
+      return this.promises[eventID](data);
+    }
+    return Promise.resolve();
+  }
+}
+
+module.exports = {
+  Parent: Parent,
+  Child: Child,
+  childProcess: childProcess,
+  EVENT: EVENT,
+};
